+++ crates/goose/src/providers/mock.rs
@@ -3,8 +3,8 @@ use async_trait::async_trait;
 use std::sync::Arc;
 use std::sync::Mutex;
 
-use crate::models::message::Message;
-use crate::models::tool::Tool;
+use crate::message::Message;
+use mcp_core::tool::Tool;
 use crate::providers::base::{Provider, Usage};
 
 /// A mock provider that returns pre-configured responses for testing
