+++ crates/goose/tests/systems.rs
@@ -3,8 +3,8 @@ use async_trait::async_trait;
 use serde_json::{json, Value};
 
 use goose::errors::{AgentError, AgentResult};
-use goose::models::content::Content;
-use goose::models::tool::{Tool, ToolCall};
+use mcp_core::content::Content;
+use mcp_core::tool::{Tool, ToolCall};
 use goose::systems::{Resource, System};
 
 /// A simple system that echoes input back to the caller
